Finish the lightweight async Rng driver.
I have not migrated OpenSK to the async driver yet.

I introduced a couple of async crates in an `async_util` module.

After some thought, I realized that having two error paths rather than routing errors through a deferred callback is the way to go. The "BUSY" error cannot be routed through the callback (as that would require a queue of unlimited length), so instead we would need to have a "cancelled" error (which signals that an ongoing operation has been cancelled). Unfortunately, it doesn't seem feasible/reasonable to cancel the kernel operation, so we would need to introduce a length-1 queue of operations in order to support the "cancelled" error. After all that, having fetch() return error codes doesn't seem so bad.